requirements.txt
altair                  4.2.0
appdirs                 1.4.4
argon2-cffi             21.3.0
argon2-cffi-bindings    21.2.0
astor                   0.8.1
asttokens               2.0.5
async-generator         1.10
attrs                   21.4.0
backcall                0.2.0
base58                  2.1.1
beautifulsoup4          4.10.0
bleach                  4.1.0
blinker                 1.4
bs4                     0.0.1
cachetools              5.0.0
certifi                 2021.10.8
cffi                    1.15.0
chardet                 4.0.0
charset-normalizer      2.0.12
click                   7.1.2
colorama                0.4.4
cryptography            36.0.1
cycler                  0.11.0
DateTime                4.4
debugpy                 1.5.1
decorator               5.1.1
defusedxml              0.7.1
docutils                0.18.1
entrypoints             0.4
et-xmlfile              1.1.0
executing               0.8.3
fonttools               4.33.3
gitdb                   4.0.9
GitPython               3.1.27
h11                     0.13.0
icrawler                0.6.6
idna                    2.10
importlib-metadata      4.11.3
ipykernel               6.9.2
ipython                 8.1.1
ipython-genutils        0.2.0
ipywidgets              7.7.0
jedi                    0.18.1
Jinja2                  3.0.3
jsonschema              4.4.0
jupyter-client          7.1.2
jupyter-core            4.9.2
jupyterlab-pygments     0.1.2
jupyterlab-widgets      1.1.0
kafka-python            2.0.2
Kivy                    2.1.0
kivy-deps.angle         0.3.2
kivy-deps.glew          0.3.1
kivy-deps.sdl2          0.4.5
Kivy-Garden             0.1.5
kiwisolver              1.4.2
libmagic                1.0
lxml                    4.8.0
Mako                    1.2.0
MarkupSafe              2.1.1
matplotlib              3.5.2
matplotlib-inline       0.1.3
mistune                 0.8.4
nbclient                0.5.13
nbconvert               6.4.4
nbformat                5.2.0
nest-asyncio            1.5.4
notebook                6.4.10
numpy                   1.22.3
opencv-python           4.5.5.64
openpyxl                3.0.9
outcome                 1.1.0
packaging               21.3
pandas                  1.4.1
pandocfilters           1.5.0
parso                   0.8.3
pickleshare             0.7.5
Pillow                  9.1.0
pip                     22.0.4
plotly                  5.6.0
prometheus-client       0.13.1
prompt-toolkit          3.0.28
protobuf                3.19.4
psutil                  5.9.0
pure-eval               0.2.2
pyarrow                 7.0.0
pycparser               2.21
pydeck                  0.7.1
pygame                  2.1.2
Pygments                2.11.2
pygubu                  0.20
pygubu-designer         0.24
pymongo                 4.1.1
Pympler                 1.0.1
pyOpenSSL               22.0.0
pyparsing               3.0.8
pypiwin32               223
pyrsistent              0.18.1
PySocks                 1.7.1
python-dateutil         2.8.2
python-magic            0.4.25
python-magic-bin        0.4.14
pytz                    2021.3
pytz-deprecation-shim   0.1.0.post0
pywin32                 303
pywinpty                2.0.5
pyzmq                   22.3.0
requests                2.25.1
selenium                3.141.0
semver                  2.13.0
Send2Trash              1.8.0
setuptools              58.1.0
six                     1.16.0
smmap                   5.0.0
sniffio                 1.2.0
sortedcontainers        2.4.0
soupsieve               2.3.1
stack-data              0.2.0
streamlit               1.7.0
tenacity                8.0.1
terminado               0.13.3
testpath                0.6.0
tk                      0.1.0
toml                    0.10.2
toolz                   0.11.2
tornado                 6.1
tqdm                    4.63.0
traitlets               5.1.1
trio                    0.20.0
trio-websocket          0.9.2
tzdata                  2021.5
tzlocal                 4.1
undetected-chromedriver 3.1.5.post4
Unidecode               1.3.4
urllib3                 1.26.4
validators              0.18.2
watchdog                2.1.6
wcwidth                 0.2.5
webdriver-manager       3.5.4
webencodings            0.5.1
websockets              10.3
wget                    3.2
widgetsnbextension      3.6.0
window                  0.0.3
wsproto                 1.1.0
XlsxWriter              3.0.3
zipp                    3.7.0
zope.interface          5.4.0